requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
redis==5.2.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
pytz==2025.1
python-dotenv
bcrypt>=4.0.1
passlib>=1.7.4
pyjwt[crypto]
python-multipart
pytest>=7.0.0
httpx>=0.23.0